WIP - review feedback: move Path assertion to org.junitpioneer.testkit.assertion
