+++ src/test/java/org/junitpioneer/testkit/assertion/PioneerPathAssert.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v20.html
  */
 
-package org.junitpioneer.internal;
+package org.junitpioneer.testkit.assertion;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Collections.singletonList;
